Move encoding package to internal sub-package
This patch moves an "encoding" package to the internal sub-package,
so it would not be accessible to the users of the main packages.
