+++ internal/encoding/encoding.go
